+++ lib/puree/configuration.rb
@@ -1,5 +1,7 @@
 module Puree
 
+  # Configuration options
+  #
   module Configuration
 
       attr_accessor :base_url, :username, :password, :basic_auth