+++ spec/project.rb
@@ -1,0 +1,108 @@
+require 'spec_helper'
+
+describe 'Project' do
+
+  def request
+    base_url = ENV['PURE_BASE_URL']
+    username = ENV['PURE_USERNAME']
+    password = ENV['PURE_PASSWORD']
+    @uuid = ENV['PURE_PROJECT_UUID']
+    @p = Puree::Project.new(base_url: base_url,
+                           username: username,
+                           password: password,
+                           basic_auth: true)
+    @metadata = @p.find uuid: @uuid
+  end
+
+  it '#new' do
+    p = Puree::Project.new
+    expect(p).to be_an_instance_of Puree::Project
+  end
+
+  describe 'data retrieval' do
+    before(:all) do
+      request
+    end
+
+    it '#find' do
+      expect(@metadata).to be_an_instance_of(Hash)
+    end
+
+    it '#acronym' do
+      expect(@p.acronym).to be_an_instance_of(String)
+    end
+
+    it '#created' do
+      expect(@p.created).to be_an_instance_of(String)
+    end
+
+    it '#description' do
+      expect(@p.description).to be_an_instance_of(String)
+    end
+
+    it '#metadata' do
+      expect(@p.metadata).to be_an_instance_of(Hash)
+    end
+
+    it '#modified' do
+      expect(@p.modified).to be_an_instance_of(String)
+    end
+
+    it '#organisation' do
+      expect(@p.organisation).to be_an_instance_of(Array)
+    end
+
+    it '#owner' do
+      expect(@p.owner).to be_an_instance_of(Hash)
+    end
+
+    it '#person' do
+      expect(@p.person).to be_an_instance_of(Hash)
+    end
+
+    it '#status' do
+      expect(@p.status).to be_an_instance_of(String)
+    end
+
+    it '#temporal' do
+      expect(@p.temporal).to be_an_instance_of(Hash)
+    end
+
+    it '#title' do
+      expect(@p.title).to be_an_instance_of(String)
+    end
+
+    it '#type' do
+      expect(@p.type).to be_an_instance_of(String)
+    end
+
+    it '#url' do
+      expect(@p.url).to be_an_instance_of(String)
+    end
+
+    it '#uuid' do
+      expect(@p.uuid).to be_an_instance_of(String)
+    end
+  end
+
+  describe 'data retrieval from file' do
+    before(:all) do
+      request
+
+      filename = "#{ENV['PURE_FILE_PATH']}project.#{@uuid}.xml"
+      File.write(filename, @p.response.body)
+
+      @metadata = @p.set_content File.read(filename)
+    end
+
+    it '#set_content' do
+      expect(@metadata).to be_an_instance_of(Hash)
+    end
+
+    it '#metadata' do
+      expect(@p.metadata).to be_an_instance_of(Hash)
+    end
+
+  end
+
+end