+++ spec/publication.rb
@@ -1,13 +1,13 @@
 require 'spec_helper'
 
-describe 'Dataset' do
+describe 'Publication' do
 
   def request
     base_url = ENV['PURE_BASE_URL']
     username = ENV['PURE_USERNAME']
     password = ENV['PURE_PASSWORD']
-    @uuid = ENV['PURE_DATASET_UUID']
-    @p = Puree::Dataset.new(base_url: base_url,
+    @uuid = ENV['PURE_PUBLICATION_UUID']
+    @p = Puree::Publication.new(base_url: base_url,
                             username: username,
                             password: password,
                             basic_auth: true)
@@ -15,8 +15,8 @@ describe 'Dataset' do
   end
 
   it '#new' do
-    p = Puree::Dataset.new
-    expect(p).to be_an_instance_of Puree::Dataset
+    p = Puree::Publication.new
+    expect(p).to be_an_instance_of Puree::Publication
   end
 
   describe 'data retrieval' do
@@ -28,16 +28,8 @@ describe 'Dataset' do
       expect(@metadata).to be_an_instance_of(Hash)
     end
 
-    it '#access' do
-      expect(@p.access).to be_an_instance_of(String)
-    end
-
-    it '#associated' do
-      expect(@p.associated).to be_an_instance_of(Array)
-    end
-
-    it '#available' do
-      expect(@p.available).to be_an_instance_of(Hash)
+    it '#category' do
+      expect(@p.category).to be_an_instance_of(String)
     end
 
     it '#created' do
@@ -52,16 +44,12 @@ describe 'Dataset' do
       expect(@p.doi).to be_an_instance_of(String)
     end
 
+    it '#event' do
+      expect(@p.event).to be_an_instance_of(Hash)
+    end
+
     it '#file' do
       expect(@p.file).to be_an_instance_of(Array)
-    end
-
-    it '#keyword' do
-      expect(@p.keyword).to be_an_instance_of(Array)
-    end
-
-    it '#link' do
-      expect(@p.link).to be_an_instance_of(Array)
     end
 
     it '#metadata' do
@@ -72,36 +60,32 @@ describe 'Dataset' do
       expect(@p.modified).to be_an_instance_of(String)
     end
 
+    it '#organisation' do
+      expect(@p.organisation).to be_an_instance_of(Array)
+    end
+
+    it '#page' do
+      expect(@p.page).to be_an_instance_of(String)
+    end
+
     it '#person' do
       expect(@p.person).to be_an_instance_of(Hash)
     end
 
-    it '#production' do
-      expect(@p.production).to be_an_instance_of(Hash)
+    it '#status' do
+      expect(@p.status).to be_an_instance_of(Array)
     end
 
-    it '#project' do
-      expect(@p.project).to be_an_instance_of(Array)
-    end
-
-    it '#publication' do
-      expect(@p.publication).to be_an_instance_of(Array)
-    end
-
-    it '#publisher' do
-      expect(@p.publisher).to be_an_instance_of(String)
-    end
-
-    it '#spatial' do
-      expect(@p.spatial).to be_an_instance_of(Array)
-    end
-
-    it '#temporal' do
-      expect(@p.temporal).to be_an_instance_of(Hash)
+    it '#subtitle' do
+      expect(@p.subtitle).to be_an_instance_of(String)
     end
 
     it '#title' do
       expect(@p.title).to be_an_instance_of(String)
+    end
+
+    it '#type' do
+      expect(@p.type).to be_an_instance_of(String)
     end
 
     it '#uuid' do
@@ -113,7 +97,7 @@ describe 'Dataset' do
     before(:all) do
       request
 
-      filename = "#{ENV['PURE_FILE_PATH']}dataset.#{@uuid}.xml"
+      filename = "#{ENV['PURE_FILE_PATH']}publication.#{@uuid}.xml"
       File.write(filename, @p.response.body)
 
       @metadata = @p.set_content File.read(filename)